add a separate mod pci_init_args
